requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
azure-storage-file-datalake>=12.9.1
djangorestframework